+++ src/socket.c
@@ -27,7 +27,7 @@
 /**
  * @brief Accepts an incoming connection attempt on a socket
  */
-int _win_accept(int s, struct sockaddr *addr, int *addrlen)
+intptr_t _win_accept(intptr_t s, struct sockaddr *addr, int *addrlen)
 {
   SOCKET r;
 
@@ -46,7 +46,7 @@ int _win_accept(int s, struct sockaddr *addr, int *addrlen)
 /**
  * @brief Accept a new connection on a socket
  */
-int _win_bind(int s, const struct sockaddr *name, int namelen)
+intptr_t _win_bind(intptr_t s, const struct sockaddr *name, int namelen)
 {
   SOCKET r;
 
@@ -62,7 +62,7 @@ int _win_bind(int s, const struct sockaddr *name, int namelen)
 /**
  * @brief Initiate a connection on a socket
  */
-int _win_connect(int s, const struct sockaddr *name, int namelen)
+int _win_connect(intptr_t s, const struct sockaddr *name, int namelen)
 {
   int iRet, iWSErr;
 
@@ -77,7 +77,7 @@ int _win_connect(int s, const struct sockaddr *name, int namelen)
 /**
  * @brief Get the name of the peer socket
  */
-int _win_getpeername(int s, struct sockaddr *name,
+int _win_getpeername(intptr_t s, struct sockaddr *name,
                 int *namelen)
 {
   int iRet = getpeername(s, name, namelen);
@@ -90,7 +90,7 @@ int _win_getpeername(int s, struct sockaddr *name,
 /**
  * @brief Get the socket name
  */
-int _win_getsockname(int s, struct sockaddr *name,
+int _win_getsockname(intptr_t s, struct sockaddr *name,
                 int *namelen)
 {
   int iRet = getsockname(s, name, namelen);
@@ -103,7 +103,7 @@ int _win_getsockname(int s, struct sockaddr *name,
 /**
  * @brief Get the socket options
  */
-int _win_getsockopt(int s, int level, int optname, char *optval, int *optlen)
+int _win_getsockopt(intptr_t s, int level, int optname, char *optval, int *optlen)
 {
   int iRet = getsockopt(s, level, optname, optval, optlen);
   if (iRet == 0 && level == SOL_SOCKET && optname == SO_ERROR)
@@ -117,7 +117,7 @@ int _win_getsockopt(int s, int level, int optname, char *optval, int *optlen)
 /**
  * @brief Listen for socket connections and limit the queue of incoming connections
  */
-int _win_listen(int s, int backlog)
+int _win_listen(intptr_t s, int backlog)
 {
   int iRet = listen(s, backlog);
 
@@ -129,7 +129,7 @@ int _win_listen(int s, int backlog)
 /**
  * @brief Receive a message from a connected socket
  */
-int _win_recv(int s, char *buf, int len, int flags)
+int _win_recv(intptr_t s, char *buf, int len, int flags)
 {
   int iRet = recv(s, buf, len, flags);
 
@@ -141,7 +141,7 @@ int _win_recv(int s, char *buf, int len, int flags)
 /**
  * @brief Receive a message from a socket
  */
-int _win_recvfrom(int s, void *buf, int len, int flags,
+int _win_recvfrom(intptr_t s, void *buf, int len, int flags,
              struct sockaddr *from, int *fromlen)
 {
   int iRet = recvfrom(s, buf, len, flags, from, fromlen);
@@ -154,7 +154,7 @@ int _win_recvfrom(int s, void *buf, int len, int flags,
 /**
  * @brief Send a message on a socket
  */
-int _win_send(int s, const char *buf, int len, int flags)
+int _win_send(intptr_t s, const char *buf, int len, int flags)
 {
   int iRet = send(s, buf, len, flags);
 
@@ -166,7 +166,7 @@ int _win_send(int s, const char *buf, int len, int flags)
 /**
  * @brief Send a message on a socket
  */
-int _win_sendto(int s, const char *buf, int len, int flags,
+int _win_sendto(intptr_t s, const char *buf, int len, int flags,
                 const struct sockaddr *to, int tolen)
 {
   int iRet = sendto(s, buf, len, flags, to, tolen);
@@ -179,7 +179,7 @@ int _win_sendto(int s, const char *buf, int len, int flags,
 /**
  * @brief Set the socket options
  */
-int _win_setsockopt(int s, int level, int optname, const void *optval,
+int _win_setsockopt(intptr_t s, int level, int optname, const void *optval,
                     int optlen)
 {
   int iRet = setsockopt(s, level, optname, (const char *) optval, optlen);
@@ -192,7 +192,7 @@ int _win_setsockopt(int s, int level, int optname, const void *optval,
 /**
  * @brief Shut down socket send and receive operations
  */
-int _win_shutdown(int s, int how)
+int _win_shutdown(intptr_t s, int how)
 {
   int iRet = shutdown(s, how);
 
@@ -204,9 +204,9 @@ int _win_shutdown(int s, int how)
 /**
  * @brief Create an endpoint for communication
  */
-int _win_socket(int af, int type, int protocol)
+intptr_t _win_socket(int af, int type, int protocol)
 {
-  int iRet;
+  intptr_t iRet;
 
   errno = 0;
 
@@ -224,7 +224,7 @@ int _win_socket(int af, int type, int protocol)
     l = 0;
     ioctlsocket(iRet, FIONBIO, &l);
 
-    __win_SetHandleType((DWORD) iRet, SOCKET_HANDLE);
+    __win_SetHandleType(iRet, SOCKET_HANDLE);
 
     return iRet;
   }
